changed uint8_t [no_]profile to uint16_t

+++ include/nigiri/loader/build_footpaths.h
@@ -4,6 +4,6 @@
 
 namespace nigiri::loader {
 
-void build_footpaths(timetable& tt, uint8_t const& no_profiles);
+void build_footpaths(timetable& tt, uint16_t const& no_profiles);
 
 }  // namespace nigiri::loader
